feat: migrate to Azure Functions
