+++ package.json
@@ -1,33 +1,31 @@
 {
-  "name": "grammybot",
+  "name": "lumebot",
   "version": "1.1.0",
+	"main": "dist/src/functions/*.js",
   "private": true,
   "scripts": {
     "deploy:pre": "bun run scripts/pre-deploy.ts",
-    "deploy:wrangler": "wrangler deploy",
-    "deploy": "run-s deploy:pre deploy:wrangler",
-    "dev:wrangler": "wrangler dev",
-    "start": "wrangler dev",
-    "dev": "bun run --watch scripts/dev.ts",
-    "test": "vitest",
-    "release": "release-it",
-    "cf-typegen": "wrangler types"
+    "build": "tsc",
+    "watch": "tsc -w",
+    "clean": "rimraf dist",
+		"dev": "run-p watch start",
+    "prestart": "npm run clean && npm run build",
+    "start": "func start",
+    "release": "release-it"
   },
   "devDependencies": {
-    "@cloudflare/vitest-pool-workers": "^0.5.2",
-    "@cloudflare/workers-types": "^4.20250109.0",
     "@types/bun": "^1.1.16",
     "@types/lodash.chunk": "^4.2.9",
     "@types/node": "^22.10.5",
     "bun": "^1.1.43",
     "release-it": "^18.1.1",
+    "rimraf": "^6.0.1",
     "type-fest": "^4.32.0",
-    "typescript": "^5.5.2",
-    "vitest": "2.1.8",
-    "wrangler": "^3.60.3"
+    "typescript": "^5.7.3"
   },
   "dependencies": {
     "@azure/data-tables": "^13.3.0",
+    "@azure/functions": "^4.6.0",
     "dayjs": "^1.11.13",
     "dotenv": "^16.4.7",
     "grammy": "github:grammyjs/grammY#support-azure-v4",
