+++ src/functions/telegramBot.ts
@@ -1,0 +1,14 @@
+import { app } from "@azure/functions";
+import { webhookCallback } from 'grammy';
+import { bootstrap } from "../bootstrap";
+
+const { bot, asyncTask } = bootstrap();
+
+app.http('telegramBot', {
+	methods: ['GET', 'POST'],
+	authLevel: 'anonymous',
+	handler: async (request, _context) => {
+		await asyncTask();
+		return webhookCallback(bot, "azure-v4")(request as any)
+	}
+});
